chore: move functional options to its own file
