+++ generic_test.go
@@ -3,15 +3,10 @@ package testcontainers
 import (
 	"context"
 	"errors"
-	"io"
 	"testing"
 
-	"github.com/docker/docker/api/types"
-	"github.com/docker/docker/api/types/filters"
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/testcontainers/testcontainers-go/exec"
 	"github.com/testcontainers/testcontainers-go/wait"
 )
 
@@ -100,93 +95,4 @@ func TestGenericReusableContainer(t *testing.T) {
 			}
 		})
 	}
-}
-
-type testExecutable struct {
-	cmds []string
-}
-
-func (t testExecutable) AsCommand() []string {
-	return t.cmds
-}
-
-func TestWithStartupCommand(t *testing.T) {
-	req := GenericContainerRequest{
-		ContainerRequest: ContainerRequest{
-			Image:      "alpine",
-			Entrypoint: []string{"tail", "-f", "/dev/null"},
-		},
-		Started: true,
-	}
-
-	testExec := testExecutable{
-		cmds: []string{"touch", "/tmp/.testcontainers"},
-	}
-
-	WithStartupCommand(testExec)(&req)
-
-	c, err := GenericContainer(context.Background(), req)
-	require.NoError(t, err)
-	defer func() {
-		err = c.Terminate(context.Background())
-		require.NoError(t, err)
-	}()
-
-	_, reader, err := c.Exec(context.Background(), []string{"ls", "/tmp/.testcontainers"}, exec.Multiplexed())
-	require.NoError(t, err)
-
-	content, err := io.ReadAll(reader)
-	require.NoError(t, err)
-	assert.Equal(t, "/tmp/.testcontainers\n", string(content))
-}
-
-func TestWithNetwork(t *testing.T) {
-	req := GenericContainerRequest{
-		ContainerRequest: ContainerRequest{},
-	}
-
-	WithNetwork("new-network", "alias")(&req)
-
-	assert.Equal(t, []string{"new-network"}, req.Networks)
-	assert.Equal(t, map[string][]string{"new-network": {"alias"}}, req.NetworkAliases)
-
-	client, err := NewDockerClientWithOpts(context.Background())
-	require.NoError(t, err)
-
-	args := filters.NewArgs()
-	args.Add("name", "new-network")
-
-	resources, err := client.NetworkList(context.Background(), types.NetworkListOptions{
-		Filters: args,
-	})
-	require.NoError(t, err)
-	assert.Len(t, resources, 1)
-
-	assert.Equal(t, "new-network", resources[0].Name)
-}
-
-func TestWithNetworkMultipleCallsWithSameNameReuseTheNetwork(t *testing.T) {
-	for int := 0; int < 100; int++ {
-		req := GenericContainerRequest{
-			ContainerRequest: ContainerRequest{},
-		}
-
-		WithNetwork("new-network", "alias")(&req)
-		assert.Equal(t, []string{"new-network"}, req.Networks)
-		assert.Equal(t, map[string][]string{"new-network": {"alias"}}, req.NetworkAliases)
-	}
-
-	client, err := NewDockerClientWithOpts(context.Background())
-	require.NoError(t, err)
-
-	args := filters.NewArgs()
-	args.Add("name", "new-network")
-
-	resources, err := client.NetworkList(context.Background(), types.NetworkListOptions{
-		Filters: args,
-	})
-	require.NoError(t, err)
-	assert.Len(t, resources, 1)
-
-	assert.Equal(t, "new-network", resources[0].Name)
 }
